Gemfile: make chef part of development group
chef v11.6.2 is installed as a global gem by the littlechef
chef-rvm recipe as defined in:

  config/lab-littlechef/roles/lab-server.json

So it doesn't need to be specified in the non-development
sections of the Gemfile.

The same rvm section in lab-server.json specifies that
ruby-2.0.0-p247@lab is the system Ruby so chef needs to
be installed to be able to run 'fix', the Python littlechef
executable to update an existing server.